Remove setTimeout and fixup tests
Seeing some performance messages about `setTimeout`s in Chrome.
This change removes the setTimeout and updates the tests so it passes.

I wrapped the second re-render in an `act` call to ensure that
the effects are flushed before checking the DOM. React is also
throwing an message about the Counter components needing `act`.

I think all of the `jest.runAllTimers()` can also be removed, but
didn't want to create a large changeset without first understanding
their need.