Add a TOC/Overview to the API section
I was looking for a page to show my devs that provides an 'API overview' for Easy-Peasy. The API detail pages are great if you already know the topic you are looking for _and_ how all the methods are related. However the section lacks a categorized 'table of contents' for those still learning. 

**This update adds a combination TOC/Overview to the API section homepage. I think it's useful for devs as a "quick reference" or "cheat sheet".** I copied most examples from the related API pages. However I fleshed some out so they are complete enough to grok how the featured method interacts with data and _other_ actions. Hopefully this provides a useful starting point.