[Bugfix] persist not working on iOS
`createPersistor` defines a `timingMethod` that won't work on iOS.

This is because this method utilizes `requestIdleCallback`, which is
defined on iOS - but never invoked 🤯

So, for iOS we have to fallback to utilize `requestAnimationFrame`.
This is done by vendoring a specific file `timingMethod.ios.js`, which
the RN bundle can import for the iOS platform.