bugfix: persist not working on iOS (#836)
`createPersistor` defines a `timingMethod` that won't work on iOS.

The `timingMethod` utilizes `requestIdleCallback` if it is defined.

It turns out that the pollyfilled requestIdleCallback for RN on iOS, requires a second parameter with a specific timeout - where as the native implementation does not. Setting the specific parameter invokes this function across all platforms.